job-matching-system/requirements/requirements-gpu.txt
fastapi
uvicorn
starlette
nest-asyncio

pdfplumber
pytesseract
docx2txt
Pillow

opencv-python-headless
deepface

transformers
accelerate

torch
torchvision
torchaudio

requests
